[FIX] sign: tests: correctly awaits the end of the test
Before this commit, when executing the test "simple rendering"
there was sometimes a crash because the blockUI div was still in the dom after the test ended.
debug mode must be 0 to maximize execution speed and therefore have a chance to reproduce.

This was because sign relies on DOM_updated event, which is not interceptable reliably by the test
and that this event fired long after the test ended according to QUnit.

After this commit and in compliance with b8afb9965d781ab0978a1e7c2db168997ce9d390, we wait until the UI
is unblocked to finish the test.

closes odoo/enterprise#13769

X-original-commit: 5586a5468656b75b9755d2c50f46278fc3170b4f
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>