[FIX] account_accountant: fix install of account installs account_accountant
Bug description:

Installing Invoicing on a database whose company is configured in the
SEPA zone, results in installing Accounting

Steps to reproduce on the Runbot:

- Start from a base enterprise database
- Go to Settings > Users & Companies > Companies and configure the
  address of MyCompany to the SEPA zone (e.g. Belgium)
- Go to Apps and install Invoicing
- The Accounting application has been installed

Explanation:

The module account_sepa depends on account_batch_payment, which depends
on account_accountant.
The post_init_hook of account, _auto_install_l10n, adds the account_sepa
module to the installation if the country_code of the company is in the
SEPA zone.
The same side-effect is observed with account_bank_statement_import_camt
which depends on account_bank_statement_import, which depends on
account_accountant.

This commit adds a new post_init_hook in the module account_accountant,
that will install these two modules if the country of the company
satisfies the SEPA conditions.

closes odoo/enterprise#13362

X-original-commit: eda6502750796a959dd42922af0cb2fc65b480a4
Related: odoo/odoo#58073
Signed-off-by: Quentin De Paoli (qdp) <qdp@openerp.com>
Signed-off-by: Paul Morelle <madprog@users.noreply.github.com>