[FIX] sale_amazon: stop anonymizing customers for missing information
Before this commit, synchronizing an Amazon sale with missing personal
information on the customer would create an anonymized partner. As the
Amazon API recently stopped to communicate some of those personal
information, partners were anonymized although some personal information
were provided, hence losing them in the process.

This commit changes the behavior of the synchronisation to stop
anonymizing partners if some of the expected information is missing and
rather create a new partner with all the available information when it
is the customer's first order, or if any information has changed since
the last order. If the buyer name if not provided, a placeholder name is
used instead.

This allows storing all the personal information provided by the API
regardless of whether some field could be missing. Additionally,
issuing an invoice now relies on the correct contact details, should
the personal information be updated after that the order is placed.

opw-2327010

closes odoo/enterprise#13817

X-original-commit: 89fdd9441c5e0acc8f76c212cf83d2607f36e22d
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>