[IMP] hr_payroll_holidays: Introduce deferred time off mechanism
Purpose
=======

Another feedback for the payroll application.

It happens that when the payslip is computed, all the employees have not
registered all their leaves. In this case, HR needs to make a modification
that will appear on the next payslip.

For the moment, when a leave day is registered afterward in the leave app,
Odoo only enables to regenerate all the payslips with the right information.

But the way the payroll works is that you do not generate a new payslip for
the employee but you only take the modification into account for the payroll
computation of the next month.

We should have an option to register all the modifications made on the work
entries after the generation of the payslip and be able to add them to the
computation of the payslip of the next month.

Specification
=============

When validation a time off, if there is already a regular payslip (no
end of the year bonus,...) which is done for that period, then set the leave
as to_defer, and prevent users to compute a new regular pay

Afterward, instead of generating the work entries for the time off (and thus
canceling the attendance at the same time), create a next activity for
the responsible, who would have to recreate a work entry at the moment he
finds the best.

Then when marking the next activity as done, this will release the leave,
and we will stop preventing users to compute another regular pay.

TaskID: 2253113