[FIX] timesheet_grid: Fix exception when starting timesheet with multiple companies
When you have multiple companies and you want to start the timer in Timesheet's app, when you choose a team, an error popup.
This occures because we are in multiple companies so the Timesheet app cannot retrieves correctly our Employee id.

This fix checks if the employee id that is linked to an account analytic line is linked to the employee that is currently connected or by one of others employees linked to the logged user.

Task id: 2326324

closes odoo/enterprise#13528

X-original-commit: 6aa93ba7eaebe2c968519f6273a566823792b67a
Related: odoo/odoo#58453
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>