[FIX] sign: Fix the kanban layout for signature template
Currently, when translation for buttons is long(mainly in french) then
it will display long button at bottom left of the card and due to that
kanban view is broken.

So in this commit, Fix the bottom left height as 50% and truncate the
text of button to display the bottom right properly instead of overlap
with each other.

closes odoo/enterprise#13584

Taskid: 2335292
X-original-commit: b47dee518873243f6eba8353fb6eb3cb420cbc48
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>