[FIX] ocn_client: missing name when registering device
Before this commit, the partner's name was used to register a device.
The problem is that this is not a mandatory field and and therefore it
can be null.

From now on, we'll use the user's name or the login if there is no name.
This change is safe because this value is only use on our own to recognize
a user/device.

We are doing this fix because of errors coming from our server.
It's usually on migration databases.

closes odoo/enterprise#13369

Forward-port-of: odoo/enterprise#13348
X-original-commit: da2d522cb56a08c3365329337c9759c27d3ee460
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>
Signed-off-by: rfr-odoo <rfr-odoo@users.noreply.github.com>