[FIX] web_studio: no access rights on custom model
Currently when the user is creating a custom model through M2O and M2M custom fields,
there are no access rights created.

So in this commit access rights are added when the user is created a custom model
through M2O and M2M custom fields and added python testcase.

closes odoo/enterprise#13925

Taskid: 2342353
X-original-commit: 3180ec142b00bd5782d805e38b22b43d70161623
Signed-off-by: Mathieu Duckerts-Antoine <Polymorphe57@users.noreply.github.com>