[IMP] hr_expense_extract: Added OCR link for Expense digitalization
Purpose of the task is, The new OCR feature in expense will be charged
via an IAP service but we do not have link to buy credit on OCR and
for the OCR services.

So in this commit, Added the 'Buy Credits' and 'View My Services'
to 'Expense' setting.

closes odoo/enterprise#14118

Taskid: 2341801
X-original-commit: 8267ce80870198dc7aa74b0a37c28d12d6645a0b
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>