[FIX] web: Improve gantt view performances
When it has a log of pills ~ 20.000, the gantt view is very slow to render.
Moving the popover to the mouse events instead of binding them to every
pill on render should drastically reduce the rendering time.

This commit improve the rendering time by 30% and also largery reduce the
DOM complexity.

closes odoo/enterprise#13508

X-original-commit: 346b949d3384f4757b6e8000b964576e6f80960d
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>