[FIX] web_gantt: Display full day instead of half day
before this commit:
In month view, if an employee take a day of time off (8am -> 4pm),
in the gantt view, only an half day will be filled.
If an employee take several days of time off, the first and the last
days will be half filled.

After this commit:
For on-day events, if the difference between the start and noon is
higher or equal to 3 hours and the difference between noon and the
end is also higher or equal to 3 hours, the event will fill all the
day.

For events over several days, if the difference between the start
and noon is higher or equal to 3 hours, the start day will be
completely filled. For the end, it's the same functioning.

closes odoo/enterprise#13527

X-original-commit: d03209161118a87e5030cdf2451a38268d5ebf4d
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: jbm-odoo <jbm-odoo@users.noreply.github.com>