[FIX] sale_renting: slow/crashing installation on existing databases.
When trying to install sale_renting app on existing databases with a lot
of existing orders, the installation may crash or take a lot of time.

This was due to some computed fields which were computed on existing
Orders & lines, taking a lot of time.

As the existing orders have no need of the rental logic, we pre-init
those fields to avoid the "useless" computations.

opw-2334712

closes odoo/enterprise#13617

X-original-commit: 69c3b94a0bfbbe44c78d01f49f723caf489d30cd
Signed-off-by: Raphael Collet (rco) <rco@openerp.com>
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>