[FIX] web_enterprise: adapt to removed o-scrollbar-overlay SCSS mixin
The mixin was removed as was only meaningfully used by the enterprise
home menu and actually encouraged to use deprecated SCSS rules. This
commit only moves its related code in the home menu scss directly. This
thus keep the inconsistency on Firefox where the overflow: overlay
rule is not understood.

closes odoo/enterprise#13532

X-original-commit: bb5d38373b462b0c6d9a041bd7dcda56b9e52f68
Related: odoo/odoo#58462
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>