[FIX] social: fix post/feed crashes due to newIDs in computed fields
Purpose of this commit is to ensure computed fields always have a value
assigned. Since compute are called by default we have more cases or compute
being called on cache / newID records. Some computation crashes as it relies
on id or because not value is assigned for cache records. In this commit
we fix several of them in social app.

Notably fixes crashes in social.post and social.post.stream where activating
studio threw OdooError.

For social.post.stream, when activating studio the computed field are called
but there is a CacheMiss in fields.py that throws a ValueError since these
fields are not set in their respective _compute functions.

For social.post, when website is also installed, there are NewIds that are
json.dumps which raises an Error.

Task ID-2345084

closes odoo/enterprise#14077

X-original-commit: 05218021c2f409411fc8686aeeedc0690cb60382
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Co-authored-by: Aurélien van Delft <avd@odoo.com>
Co-authored-by: Thibault Delavallée <tde@odoo.com>