[FIX] web: Avoid home menu scrollbar flicker
The mixin "o-scrollbar-overlay" has been removed since it layed on
non-standard CSS properties to make the scrollbar overlap on the
content.

The initial intent was to have the scrollbar displayed on top of its
related content using a simple CSS-only solution. Unfortunately such a
property does not exist in the current standards.

As such, this fix consists of a small hack in the home menu component
meant to compensate the scrollbar width with an equally large padding
left calculated on each visual update.

/!\ as the standards evolve we may someday be able to use a one-line
css property to achieve this. If/when this happens this fix must be
adapted accordingly.

closes odoo/enterprise#14049

X-original-commit: 4e9d2cb8907f081e8ef38d2397ebf8d89d042e14
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>