[FIX] account_asset: missing partner when asset auto-creation
- Set an account A with "Automatic Asset Creation" set
- Create an invoice
- Add a product, set the account to A
- Post the invoice

When the depreciation board is computed, the partner of the invoice is
not carried out to the journal entries and items.

To retrieve the partner, we use the `original_move_line_ids` field to
get all partners. If there is no more than one (which should be the
case most of the time), set it on the journal entry and items.

opw-2335929

closes odoo/enterprise#13510

closes odoo/enterprise#13534

closes odoo/enterprise#13690

X-original-commit: a76a603c862036764766177c9f26307469ffda28
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>