[REF] stock_barcode: replace shortcut
Some barcode scanner similate the tipping of the barcode's chars, which
could to lead to trigger the shortcut we manually set on the buttons in
the Barcode App. The purpose of this commit is to remove these shortcuts
and use instead the Odoo default shortcuts (the ones who appears when
the user press Alt key).

task-2333834

closes odoo/enterprise#13636

X-original-commit: 42fec3f283deb9d99a908bbe00c96183e6f8e231
Related: odoo/upgrade#1818
Signed-off-by: William Henrotin <Whenrow@users.noreply.github.com>
Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>