[IMP] l10n_be_hr_payroll: compute paid time off for the next year
Before this commit, we computed the amount of the paid time off according
to the number of worked days that employee has made during the year.
Now, we compute the correct number of legal leaves according to the
working schedule in the contract(s) of each employee during the current
year.

Review allocated days according working schedule
------------------------------------------------

In the Paid Time Off Allocation wizard, we also need to compute the
legal leaves according the current contract for the next period that
the employees have.

To find the contract for the next year, we need to search also in the
incoming contract, because if we compute the legal leaves at the end of
the period. Then it is possible that an employee finished his contract at
the end of year and has another contract for the begining of the next
year (next period).

We need to add a field named "max_leaves_allocated" in
'hr.leave.allocation' model to retain the maximum time offs that we can
allocate to an employee for the next year when he changes his working
schedule. Indeed, if in the current period, we compute 18 legal leaves
for the following period of an employee, if his contract at the next
period specifies that he works at mid-time, we need to allocate the half
of legal leaves computed but we must retain the computed amount for the
next period to not allocate too many legal leaves when the employee
would want to change his working schedule.

Remove the worked days field in Paid Time Off Allocation wizard
---------------------------------------------------------------

Since we now compute the legal leaves with the working schedule in the
contracts during the period, the worked days field is useless to keep
in this wizard.

Change date range of period to year
-----------------------------------

The using of the Paid Time Off Allocation wizard is only when we need to
allocate Paid Time Off for each employee in belgian.
Therefore, the selection of a year of the reference period is more
user-friendly then selected a date range.

task-2314223