[FIX] account_consolidation: allow deleting a consolidation chart if it is still used by groups, deleting the groups as well
Consolidation chart object is central to the the consolidation module. If it is deleted, all the depending objects also must be, including groups.

The only exception to this is consolidation entries (whose model is named consolidation.journal). We don't want to cascade on them when deleting a chart, as it could be and error and would lead to data loss. If the user really wants to delete the chart in this case, she'll have to delete all the entries first.

X-original-commit: c7959c5e1bbc770d88522226aa34c6692819b992