[IMP] l10n_au_keypay: add module
New module to synchronize Australian payroll with odoo

The settings are at the bottom of the accounting section,
you need an Australian company to see them.

You can setup one api key and the url of the provider
as a global seting while activating the integration per
company, you will need it's business id to do so.

task-2280848

closes odoo/enterprise#15933

X-original-commit: f251bf1152b4f7f0410963469fac34d6470433ee
Signed-off-by: Quentin De Paoli (qdp) <qdp@openerp.com>