[ADD] sale_subscription_timesheet: subscription do not get analytic account
Create a sale order for a subscription product which is set to
'Create a task in sales order's project' (Project).
Save and confirm the Sale order.

The created subscription will not have the analytic account
as the sale order because the aa is created after the subscription

Adding a bridge module to synchronize the data manually

opw-2341121

closes odoo/enterprise#13622

closes odoo/enterprise#13727

Signed-off-by: agr-odoo <agr-odoo@users.noreply.github.com>
Signed-off-by: Xavier Dollé (xdo) <xdo@odoo.com>