[FIX] account_consolidation: fix company record rules
If, for a reason, a consolidation.chart does not have any company_ids,
it is not possible to see its dependents (groups, accounts, periods),
and if there are, it is not possible to delete the chart because of the
foreign keys.

This leads to an inexplicable situation from the user point of view,
where the consolidation looks empty but cannot be deleted.

This commit fixes the problem by allowing the visibility on any
consolidation item if its chart has no company at all, which is the
traditional behavior in Odoo for multi-company fields.

closes odoo/enterprise#13706

X-original-commit: 6cf0e8253623d82af006392f55a0eb2c58f350e5
Signed-off-by: Paul Morelle <madprog@users.noreply.github.com>