[FIX] marketing_automation: cannot change mail template
Steps to reproduce the bug:
- Create a new campaign in Marketing Automation
- Add a new activity
- Launch a test
- Run the test (click on the "play" icon)
- Go back in the campaign and edit it
- Pick a new Mail Template and save
- Launch a test again
- Run the test

Bug:
Traceback: a record already exists for the PG unique constraint.

Cause:
A new field is added when installing mass_mailing, which is a dependency of marketing_automation and the link_tracker uniqueness constraint is not updated.
When creating a new link_tracker, it looks for existing records with mass_mailing_id in its search criteria.
At the moment, this bug has only been triggered in test mode, hence this workaround.

Closes odoo/odoo#57144

opw:2335523

closes odoo/enterprise#13694

X-original-commit: d415831601b6ffc216c1b362974c0e6203473cc3
Signed-off-by: backspac <backspac@users.noreply.github.com>