[FIX] Planning: Fix allocated % modifies start & end
In planning module, when we create & save a new shift with other start & end times than the default one and then we open it & change the allocated percentage, the start & end times are changed.
In fact, start & end times are adjusted with our hour interval but it should not.

This commit checks if there was an existing start & end datetimes and if there is, it will not update datetimes

Task id: 2333145

closes odoo/enterprise#13822

X-original-commit: effa6eb2a688d273aa89735f5c0f99c5be580b51
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>