[FIX] delivery_fedex: request to get rates in preferrred currency
Without the "RateRequestTypes" set to "PREFERRED", the service only
returns the rate in the origin's currency.  This causes an error
when we attempt to read the result (in
ProviderFedex.fedex_rate_shipment).  With RateRequestTypes = PREFERRED
Fedex will return the rates in both currencies (the one requested + the
origin), so there should be no impact on existing instance.
At the same time, the code to extract the price from the Fedex response
was made a little more defensive (the code that was reading USD was
actually wrong, because it used the wrong currency to convert, but this
should normally not be reached anyway with the RateRequestTypes set to
PREFERRED).  In case all else fails and no known currency is returned,
we'll log a message and return 0 rather than throwing an error.

Steps to reproduce (on 12.0 or 13.0):

- Configure a warehouse in UK.  Make sure to enter a valid UK address
 - I used 10 Downing Street, London SW1A 2AA
- Configure Fedex International shipping method
- Company currency is not relevant - I used USD.  It is not necessary for multi currency to be enabled.
- Make a Sales order in USD.  Select the UK warehouse as origin, and select a valid US address as destination.
- Add Shipping and click Get Rate

opw-2343445

closes odoo/enterprise#13805

X-original-commit: 1c19d90cb67cd7f970e9481024b548597a285272
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>
Signed-off-by: Nicolas Galler <nicocrm@users.noreply.github.com>