[IMP] l10n_be_hr_payroll: reevaluate the right to time off
reevaluate the time off when a credit time is made for an employee
------------------------------------------------------------------

When the employee needs to a credit to change his working schedule, we
need to reevaluate the right to time off at each change of working
schedule.

reevaluate the time off when exit credit time
---------------------------------------------

When the user removes the credit time for an employee to create a new
contract at full time, we need to reevaluate the paid time off to
allocate.

compute the old work time rate
------------------------------

We need to compute the old work time rate of the employee because we
must recompute the correct work time rate to compute correctly the paid
time off.

task-2314223