[FIX] account_asset: last day of purchase
fw-port of 1bb802ecfdc9cdc8f72e092d79a53bc3de3cd549

- Create an asset with:
Depreciation Dates: Based on Last Day of Purchase Period
Prorata Temporis: true
Number of Depreciations: 12
Number of Months in a Period: 1
- Compute the depreciation

The depreciation day should be the last day of the month, each month.
However, it is not the case and ultimately becomes the 28th or 29th.

The use case is already taken into account for when Prorata Temporis is
false. There is no good reason to not do it in this case.

opw-2341728

closes odoo/enterprise#13737

X-original-commit: 6c607375ed2707568b5544b61777ece51ed6927d
Signed-off-by: agr-odoo <agr-odoo@users.noreply.github.com>
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>