[FIX] documents: apply mobile customization on list view
In Documents the Kanban view was customized to accommodate usage on
smaller screens. But when the List view was introduced, those
customization were not applied.

This commit applies the same customization as the Kanban to the List
view.

Note: due to this fix being on a stable branch files cannot be renamed.
Their names still contains "kanban" even if they apply on both Kanban
and List views ; this will be fixed on master branch.

Task ID: 2350354

closes odoo/enterprise#14053

X-original-commit: b7eab93e670477f6dd146b1d7650b754ebdda0d7
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>
Signed-off-by: Pierre Paridans <pparidans@users.noreply.github.com>