[FIX] l10n_co_edi_ubl_2_1: display the currency rate with 2 decimals
The _convert function will round the result using the rounding field
on the resulting currency. In this case the resulting currency will
always be COP, since this module is meant for Colombian companies. COP
has no cents, all amounts are integers. As a result currency_rate
always ends up being "x.00".

This solves that by manually calculating the value without the
_convert method. Note that the "%.2f" format specifier doesn't
truncate the value. It correctly rounds, similar to using round(x, 2).

opw-2350026

closes odoo/enterprise#14088

X-original-commit: 4e3aba7f40655e3db3f7a299bbc0e13e21cc5cd5
Signed-off-by: jorenvo <jorenvo@users.noreply.github.com>