[FIX] planning: show monthly view if user sends monthly planning
When a user would send monthly planning, if an employee had only one shift and that shift was in the last three weeks of the month, he would receive a view of the first (and empty) week.
This was caused by the code always trying to send the span sent by the user but showing weekly view if the employee's shift only spanned one week.
The fix was to base the view type (week/month) based on what the user sent instead of the employee's shift.

task-2326229