[FIX] account_reports: fix infinite recursion when folding financial report line
An infinite recursion happened in case a financial report line was using a groupby, and one of its children had the same id as its parent (which was possible because line ids were built from different models), when trying to fold the line. We now make the ids of the lines created by the groupby unique, and modify the caret options available on such grouped lines accordingly (that is, only the ones applicable to account.account, by design).

OPW 2301756

[FIX] account_reports: manage annotations correctly on groupby-created lines in case a group is repeated across multiple report lines.

This second bug also happened because those groups share the same id. The annotation made for one were shared between all of them. When making an annotation, we expect it to be displayed only on the line we click ("annotate") for, not to spread it on other lines sharing some characteristic with it (for example, the same account or partner).

closes odoo/enterprise#13931

X-original-commit: d1f860670541b4d36bf2a078620cd368353dbe74
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>