[FIX] website_helpdesk_forum: Fix the display of False on the forum
Currently, when go to helpdesk > helpdesk teams > enable 'help center'
create a ticket (leave the description empty) > share on the forum
It will display the 'False' on the forum post.
Because when form.post is created at that time it will convert the
description to html and False value is converted to html due to
that it was not going to OR condition.

So in this commit, check for the description and if description exist
then and then convert it to html OR pass the blank string.

closes odoo/enterprise#14141

Taskid: 2345334
X-original-commit: fee40d61d70baf66d9769e34e5ccc960780f1923
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>