[IMP] hr_payroll_account: Split method that generates the journal entry
The method that generates the journal entry for the payroll is many
complex, and is not easy to inherit to change some values if it´s
necessary.

Now, were split in 2 new methods:
- To get the data to set on the journal items, to en sure that could
be added any other value in the line if is caller from other module.
- To allow change the way as verify if the debit/credit line exists.

With this, is more easy extends the module, and, avoid overwrite all the
method if it´s necessary.

closes odoo/enterprise#14426

X-original-commit: bcf4cdb57fdeb2f714894e12b9d32ae5ef8b6c16
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>