[FIX] social_linkedin: correctly mock method to avoid HTTP call
This commit fixes an issue with the social_linkedin tests where the test case
would make an HTTP request call to the linkedin API during the social.account
creation that is used for the test.

Obviously we want to avoid HTTP calls during automated tests, as it's also
completely unnecessary for the test purpose we just mock the
"_compute_statistics" method to do nothing, which in turn resolves our issue.

Task 2351507

closes odoo/enterprise#13859

X-original-commit: 5e894550717b8526f9262f9b06d429ef655d21c0
Signed-off-by: awa-odoo <awa-odoo@users.noreply.github.com>