[FIX] delivery_easypost: fix weight when returning of a return
- Install delivery_easypost
- Go to Sales > Configuration > Shipping Methods and create a shipping method:
  * Name: Easypost Fedex
  * Provider: Easypost
  * Carrier Type: FedEx
  * Default Package Type: FedExBox
  * Default Service Level: FEDEX_2_DAY
  * (Test Environment)
- Go to Sales > Products and create a storable product with a weight of 0.1 (i.e. Product X)
- Update the quantity (i.e. 100)
- Go to Sales > Orders > Customers and create a customer:
  * Customer X
  * Company Address: 425 Recker HwyAuburndale
  *                  Florida (US)
  *                  33823
  *                  United States
  * Phone: 00337123456
- Create a Quotation:
  * Customer: Customer X
  * Order Lines: Product X (quantity: 10)
  * Add Shipping: Easypost Fedex and get rate
- Confirm Quotation
- Open delivery via Delivery smart button
- Check and set (if not already set) Carrier to "Easypost Fedex" in Additional Info tab
- Validate
- Create a return (via return button), set Carrier (Easypost Fedex) and validate
- Create a return of the previous return (via return button), set Carrier (Easypost Fedex) and validate
An error is raised: "Easypost returned an error: Validation failed: Weight must be greater than 0"

The weight is computed from `qty_done` of the `stock.move.line` for a normal delivery (outgoing) and
from `product_qty` for a return delivery (incoming).
The computed field `is_return_picking` defines if a picking is a return or not.
The issue comes from the fact that `is_return_picking` is True if it is a return of another picking and
and flags the return of a return as a return, when it should not as it is an outgoing picking.
Therefore weight is computed from `product_qty` (which is 0) instead of `qty_done`.

Instead of basing the weight computation on `is_return_picking` field, it should be done on the picking type.

opw-2312425

closes odoo/enterprise#13864

X-original-commit: 153ed8fb5b43e98d254c71f3f0f1e0a545c24e5e
Signed-off-by: Anh Thao PHAM <kitan191@users.noreply.github.com>