[FIX] documents_spreadsheet: Fetch updated cache when reinserting pivot
Steps to reproduce:
- insert a pivot
- update a record aggregated in the pivot such that it falls into anoter group
- re-insert the pivot

=> the record does not appear in its new group.

To re-insert the pivot we need to refetch the cache to get the new structure.
As a performance optimization, the previous cache was kept and re-used to avoid
having to re-apply (potential) filters.
This is wrong since some measure might have also changed and the user expect to
have up-to-date data in a pivot he just inserted.

Unfortunatly, this means potentially fetching the cache twice:
1) without filters to have the full structure (and get up-to-date data)
2) with the filters

Task 2354996

closes odoo/enterprise#14716

X-original-commit: 15b90462dff3901cfe57a4279e2a39538328e1da
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>