[IMP] sign: Improve the email sent view on sign document
Currently, When we click on 'sign now' or directly sign with shared
link it display the resend button even if we have not sned any mail
but it's not true.

So in this commit, we display the 'send' button in the sign document
if we never sent a mail to other signatory or we will display 'Resend'
if already mail is sent.
Also added buttons for send/resned on list view of sign item.

closes odoo/enterprise#13880

Taskid: 2275857
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>