[FIX] web_gantt: no flickering with large popovers
Before this commit, when the content of a gantt pill popover was
very long, the popover couldn't open, and the pill flickered.

This commit changes the position of the popover from 'top' to
'auto', such that if the content is too long and there is no space
in top then it will adjust itself and open it wherever there is
the most space.

task-2348608

closes odoo/enterprise#15418

X-original-commit: 8cee8ae1219e082eb03b93838332c68ccbec9183
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>