[IMP] l10n_be_hr_payroll: add unit test for allocating paid time off
This commit add 2 tests for allocating paid time off. In the first, we
check if we correctly compute the right amount of leaves for the next
year. In the next test, we check if we allocate the correct amount of
paid time off (leaves) based on the working schedule of the employee for
the next year.