[FIX] documents: Trackback on dragging document on workspace category
Currently, while dragging the document from the central document page
to another workspace label it throws trackback in Firefox because
When you have span/div with droppable attribute and that span/div has
text inside, if you drag another element over span/div it correctly fires
dragenter event but if you drag over the text inside that span/div it
fires dragleave - even if the text is inside droppable span/div.

So in this commit, we put style on span with pointer-events: none;
so everything works fine.

closes odoo/enterprise#14813

Taskid: 2349799
X-original-commit: 6104a44b32d7024d12330b929c3d40cf908051c8
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>