[FIX] account: in loadData, pass the correct value to _formatLine
When lines are loaded in the reconciliation widget, get_bank_statement_data
and get_bank_statement_line_data both return the same object containing e.g.
{st_line: [], reconciliation_proposition: [], lines: []}.
Then, the value of 'lines' should be passed to _formatLine. However
in loadData, which is called after reconciling an entry, and if an additional
line should be loaded in the reconciliation widget, the value passed was
incorrect. Before saas-12.3, res['lines'] was passed to _formatLine. We
do the same change to ensure the correct value is passed to _formatLine.

opw 2299574

closes odoo/enterprise#13961

Forward-port-of: https://github.com/odoo/odoo/pull/59428
X-original-commit: 6c1a1169f8e589ed019c3aa905da28844e7cbb9b
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>