[FIX] web_gantt: keep same time when dragging to/from DST period
Behavior prior to the fix:

If I, in the Brussels timezone, move or copy a task (on the planner) from
10/20/20 to 10/27/20, and it was originally scheduled at 08:00 CEST, the
schedule will change to 07:00 CEST (still 06:00 GMT in both cases)

Behavior after the fix:

The system uses the web user's timezone (as identified by their
browser) to perform the time adjustment calculation. The second occurence
will thus be on 10/27 08:00 local time (CET) (so 07:00 GMT).

NOTE: Another commit will follow for opw-2349735 to implement the backend
change necessary to the planning module for recurrency and copy period
functions, as well as the performance issue identified on that ticket.

opw-2349735

closes odoo/enterprise#14334

X-original-commit: d8d165ed19f89cd330dd03235a974e5818d3ce66
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>
Signed-off-by: Nicolas Galler <nicocrm@users.noreply.github.com>