[FIX] planning: compute recurrence in browser's timezone
Behavior prior to the fix:

- If I, in the Brussels timezone, schedule a recurring task for an employee
on 10/20/2020 08:00 (local time = CEST), and have it recur weekly,
the next occurence will be on 10/27/2020 07:00 (local time = CET). In
both cases the time is 06:00 GMT, but the employee is interested in
local time.
- Likewise, when using the "Copy previous week" action on the planner,
the GMT timestamps are preserved, causing a shift in the local time

Behavior after the fix:

- The system uses the web user's timezone (as identified by their
browser) to perform the recurrency calculation. The second occurence
will thus be on 10/27 08:00 local time (07:00 GMT).

NOTES:

- when entering a schedule, the timezone is currently computed based
on the local timezone of the user (the one set on his computer). It
may be preferrable to use the employee's timezone. Such a change is not
part of the scope of the fix since it is a behavior change (worth
noting also that Odoo generally uses the browser's timezone rather than
the user's timezone)
- the reporting of the time in the web interface is not consistent (if
my timezone is different from that in which the event was entered, the
time may be reported differently on the planner UI and in the Planning
edit popup). The fix does not intend to modify this behavior either.
- the test on period duplication (TestPeriodDuplication) which had been
commented as part of commit 3e957aa
(when the recurrency logic was modified) was uncommented and corrected
as part of the fix, since an additional test was needed to check the
behavior of the period duplication function across different timezones
- another commit follows to address the performance concern noted on
ticket opw-2349735

See also https://github.com/odoo/enterprise/pull/13968 (fix for the copy
/ move functions on the gantt frontend)

opw-2349735

closes odoo/enterprise#14360

X-original-commit: ce99c36ad001b34ccd3bbfa3f8581da763d919a0
Signed-off-by: Nicolas Galler <nicocrm@users.noreply.github.com>
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>