[FIX] planning: fix query in compute_overlap_slot_count
Behavior prior to this fix:

With a large number of planning_slot entries, the
compute_overlap_slot_count method takes a long time to run.

Behavior after this fix:

The query was limited to retrieve only the ids for the records that we
are computing (in fact the parameter was already there, so it's likely
an oversight that it was not included in the query).  Execution time on
the customer database was reduced from 4 seconds to 0.2 second.

opw-2349735

closes odoo/enterprise#13988

X-original-commit: 6aae9a214fda01fb8ef441e572193655c1f340cb
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>