[IMP] event_barcode: clean ACL and add a new group "Registration Desk"
Purpose
=======
Clean the ACLs related to the Event application.

Add a new group to manage the registration in the entrance of an event.
This group should not be able to "break" things in Event but should be
able to create and manage the registrations.

Specification
=============
This new group should be able to use event barcode.

Task 2204364
COM odoo/odoo/pull/57022
ENT odoo/enterprise/pull/13984
UPG odoo/upgrade/pull/1897

closes odoo/enterprise#13984

Related: odoo/upgrade#1897
Related: odoo/odoo#57022
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>