[FIX] documents: preserve mime type for uploaded Word documents
Behavior before the fix:

- When uploading an office 2007 document (.xlsx / .docx) as a regular
(non admin) user, the preview shows it as text, and the download option
by default saves it as a text document (in certain browsers, at least).
This stems from an over-aggressive check on mime type (anything that
contains 'xml' **anywhere** in the mime type is taken to be XML, but the
mime type for an Office document is
`application/vnd.openxmlformats-officedocument.wordprocessingml.document`
so it falls in the XML case.

Behavior after the fix:

- The mime type is preserved
- Mime type for XML-like documents (including HTML and SVG) continues to
be validated

opw-2352712

closes odoo/enterprise#14168

X-original-commit: b32869a04ed32145bd5a12b8d9bbb60a3ae07ea2
Related: odoo/odoo#60164
Signed-off-by: Nicolas Galler <nicocrm@users.noreply.github.com>
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>