[FIX] sale_subscription: fix missing invoices from Subscriptions in Documents
- Install Accounting, Subscriptions and Documents
- Go to Documents > Configuration > Settings and activate Accounting:
  * Workspace: Finance
  * Journals: Customer Invoices - Finance
- Go to Subscriptions > Configuration > Subscription Templates
- Create or edit a Template (i.e. Mon - Monthly Subscription) with "Send" for "Create Invoice"
- Create a Subscription:
  * Subscription Template: Mon - Monthly Subscription
  * Subscription Lines: [Any a product]
- Switch Subscription state from "Draft" to "In Progress"
- Click on "Generate Invoice" button
If it is the first generated invoice, it will be opened automatically and there will be
no issue. If that is the case, generate another invoice and do not open it.
- Go to Documents > Documents and select "Finance" Workspace
The generated invoice from Subscription will not appear as long as it will not be opened.

The Document is created from "message_main_attachment_id" field of invoice, but this field is not
set when invoice is created and sent in Subscriptions.

opw-2345618

closes odoo/enterprise#14003

closes odoo/enterprise#14082

X-original-commit: 7e075cc6f16a00a3d6a809cdf0998f3d666afe24
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>
Signed-off-by: Anh Thao PHAM <kitan191@users.noreply.github.com>