[FIX] account_reports: remove currency total on general ledger
When a company made transactions in multiple currencies, on the general
ledger when selecting a journal, the amount currency of the journal
items are shown. Totals for the journal items are also calculated :
total of debit, total of credit, the final balance and total of amount
currency.

The problem with the total of amount currency, is that the company can
make transactions in multiple currencies, and sum this amounts in
multiple currencies don't have much sense, its like adding apples and
oranges. Also, the currency symbol of this total is the one of the
currency of the account, which, as explained before, can be different
from the amount currency of the journal items.

opw-2346057

closes odoo/enterprise#14199

X-original-commit: 18d5a697ca7e1424acd86f6efe8cb8caaf4efe23
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>