[FIX] *: adapt business code to function-redefined error
This commit adapts the business code in which
class/module/function/method redefinition took place so that it no
longer happens and the pylint test passes.

This commit also fixes a bug in the web_studio test
`test_view_normalization_37` test that was using `div` elements inside
an embedded tree view which is not allowed, instead we use `button`
elements for the test which are allowed and make the test pass, note
that this not affect the purpose of the test which is to test that
the correct elements are selected when there are multiple elements with
the same name.

closes odoo/enterprise#14090

Related: odoo/odoo#60044
Signed-off-by: Adrian Torres (adt) <adt@odoo.com>